refactor: add emotion & fix [#10]
