+++ .babelrc
@@ -1,13 +1,14 @@
 {
-  "presets": ["next/babel"],
-  "plugins": [
+  "presets": [
     [
-      "styled-components",
+      "next/babel",
       {
-        "ssr": true,
-        "displayName": true,
-        "preprocess": false
+        "preset-react": {
+          "runtime": "automatic",
+          "importSource": "@emotion/react"
+        }
       }
     ]
-  ]
+  ],
+  "plugins": ["@emotion/babel-plugin"]
 }
