+++ package.json
@@ -11,10 +11,13 @@
     "prod": "next build && next export"
   },
   "dependencies": {
+    "@emotion/babel-preset-css-prop": "^11.2.0",
+    "@emotion/react": "^11.1.5",
+    "@emotion/styled": "^11.3.0",
     "axios": "^0.21.1",
     "babel-plugin-styled-components": "^1.12.0",
     "moment": "^2.29.1",
-    "next": "latest",
+    "next": "^10.2.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
     "react-helmet": "^6.1.0",
@@ -22,6 +25,8 @@
     "styled-components": "^5.2.3"
   },
   "devDependencies": {
+    "@babel/core": "^7.14.0",
+    "@emotion/babel-plugin": "^11.3.0",
     "@types/node": "^12.12.21",
     "@types/react": "^17.0.2",
     "@types/react-dom": "^17.0.1",
