+++ src/components/header/header.styled.tsx
@@ -1,5 +1,5 @@
 //@ts-nocheck
-import { theme } from '@components/utils/styles/theme';
+import { theme } from 'src/constants/theme';
 import { StyleType } from '@interfaces/style';
 import styled from 'styled-components';
