+++ src/components/header/header.tsx
@@ -1,21 +1,22 @@
+/** @jsxImportSource @emotion/react */
+import { css, SerializedStyles } from '@emotion/react';
+
 /**
  * @description 어플리케이션의 헤더
  *              path에 따라 커스텀된 헤더 반환
  */
 
-import React from 'react';
 // import { HeaderType } from '@interfaces/app';
 // import { HeaderContainer } from './header.styled';
-import { theme } from '@components/utils/styles/theme';
+import { theme } from 'src/constants/theme';
 import MainHeaderContent from './MainHeader';
-import { CSSProperties } from 'styled-components';
 
 interface HeaderProps {
   pathname: string;
 }
 
 interface HeaderStyle {
-  [key: string]: CSSProperties;
+  [key: string]: SerializedStyles;
 }
 
 const HeaderContents: any = {
@@ -24,21 +25,21 @@ const HeaderContents: any = {
 };
 
 const Header = ({ pathname, ...rest }: HeaderProps) => {
-  return <div style={styles[pathname]}>{HeaderContents[pathname](rest)}</div>;
+  return <div css={styles[pathname]}>{HeaderContents[pathname](rest)}</div>;
 };
 
 export default Header;
 
 const styles: HeaderStyle = {
-  '/': {
-    width: theme.padding.width,
-    height: '5vh',
-    display: 'flex',
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-    alignItems: 'center',
-    padding: theme.padding.base,
-    backgroundColor: theme.color.main,
-    color: theme.color.white,
-  },
+  '/': css`
+    width: ${theme.padding.width};
+    height: 5vh;
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-end;
+    align-items: center;
+    padding: ${theme.padding.base};
+    background-color: ${theme.color.main};
+    color: ${theme.color.white};
+  `,
 };
