+++ src/components/header/index.tsx
@@ -1,1 +1,1 @@
-export { default } from './header';
+export { default } from './Header';
