+++ src/components/main/Main.presenter.tsx
@@ -7,7 +7,7 @@ import * as ms from './Main.styled';
 import Feed from './items/Feed.container';
 import { ScrollAnimationItem } from '@components/main/items/ScrollAnimationItem';
 import { LTitle, TitleContainer } from '@components/utils/styles/title.styled';
-import { theme } from '@components/utils/styles/theme';
+import { theme } from 'src/constants/theme';
 
 interface MarginType {
   [idx: string]: {
