+++ src/components/main/Main.styled.ts
@@ -1,4 +1,4 @@
-import { theme } from '@components/utils/styles/theme';
+import { theme } from 'src/constants/theme';
 import { StyleType } from '@interfaces/style';
 import styled from 'styled-components';
