+++ src/components/main/items/Feed.presenter.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import { BsCaretRightFill } from 'react-icons/bs';
 import { ItemContainer } from '@components/main/items/items.styled';
 import { TitleContainer, XLTitle } from '@components/utils/styles/title.styled';
-import { theme } from '@components/utils/styles/theme';
+import { theme } from 'src/constants/theme';
 import * as ms from '../Main.styled';
 
 import { FeedPropsType } from '@interfaces/components';
