+++ src/components/utils/styles/container.styled.ts
@@ -1,4 +1,4 @@
-import styled from 'styled-components';
+import styled from '@emotion/styled';
 
 export const OuterContainer = styled.div`
   width: 100vw;
